+++ common/src/main/java/com/human/common/registry/init/creative_mode_tab/initializer/CreativeModeTabUtil.java
@@ -1,6 +1,7 @@
 package com.human.common.registry.init.creative_mode_tab.initializer;
 
 import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.ItemLike;
 
 import java.util.function.Supplier;
@@ -13,5 +14,9 @@ public class CreativeModeTabUtil {
 
     /* package-private */ static void accept(CreativeModeTab.Output output, ItemLike itemLike) {
         output.accept(itemLike);
+    }
+
+    /* package-private */ static void accept(CreativeModeTab.Output output, ItemStack itemStack) {
+        output.accept(itemStack);
     }
 }